Create Coupon_Collector.py
Code for a problem known as Coupon Collector.

Create a program that models the coupon collector problem. In this problem, you have a set of 'n' distinct coupons, and each time you collect a coupon, it's equally likely to be any of the 'n' types. Calculate the expected number of coupons you need to collect to complete the set.